backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
supabase>=2.0.0
PyPDF2>=3.0.0
requests>=2.32.0
elevenlabs>=1.0.0
python-multipart>=0.0.12
aiofiles>=24.1.0
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.0.0
